requirements.txt
# Python 3.12.2 Compatible Requirements
# Base Flask Dependencies
Flask==3.0.2
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.27
pyodbc==5.0.1
pymssql==2.2.11 

# Authentication & Security
PyJWT==2.8.0
python-dotenv==1.0.1

# Production Server
gunicorn==21.2.0
gevent==24.2.1

# Monitoring & Logging
prometheus-flask-exporter==0.23.0
python-json-logger==2.0.7

# Caching
redis==5.0.1
Flask-Caching==2.1.0

# Performance & Utils
psutil==5.9.8